Move scheduler components into the directory
