+++ src/components/shared/scheduler/SchedulerText.vue
